+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes
 TypeTree 2.10-git
 ----------------
 
+- `HybridTreePath` functions now have `[[nodiscard]]` attribute
 - `HybridTreePath` gained several multi-index utilities:
   `max_size`, `join`, `reverse`, and `accumulate_[front|back]`.
 - Add user-defined literal `_tp` to generate `HybridTreePath`